components/pep/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
psycopg2-binary==2.9.9
